+++ react/components/ReviewBlock.tsx
@@ -433,8 +433,6 @@ const ReviewBlock: FunctionComponent<WrappedComponentProps & any> = ({
 
   const createSchema = (columnsToBeHidden: string[]) => {
     if (columnsToBeHidden.indexOf('line') === -1) {
-      let count = 0
-
       tableSchema.properties.line = {
         type: 'object',
         title: intl.formatMessage({
@@ -442,10 +440,8 @@ const ReviewBlock: FunctionComponent<WrappedComponentProps & any> = ({
         }),
         width: 50,
         // eslint-disable-next-line react/display-name
-        cellRenderer: () => {
-          count++
-
-          return <div>{refidLoading ? '-' : count}</div>
+        cellRenderer: ({ rowData }: any) => {
+          return <div> {refidLoading ? '-' : rowData.index + 1}</div>
         },
       }
     }